backend/requirements.txt
Django==4.2.19
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# Úteis/seguras para o projeto (recomendado manter)
Pillow==10.4.0            # uploads/imagens (thumbnails, etc.)
python-dotenv==1.0.1      # carregar variáveis de ambiente de .env

# Produção (Linux) - opcional
gunicorn==21.2.0          # WSGI server
whitenoise==6.7.0         # servir staticfiles sem Nginx (apps simples)

# Banco Postgres (opcional; em dev usamos SQLite)
psycopg2-binary==2.9.9

requests>=2.31.0
